updated some docstrings
